:truck: chore: refactored settings state
