+++ app/src/main/java/com/shub39/grit/core/data/backup/restore/RestoreImpl.kt
@@ -23,11 +23,11 @@ import com.shub39.grit.core.data.backup.toHabit
 import com.shub39.grit.core.data.backup.toHabitStatus
 import com.shub39.grit.core.data.backup.toTask
 import com.shub39.grit.core.domain.AlarmScheduler
-import com.shub39.grit.core.domain.backup.RestoreFailedException
-import com.shub39.grit.core.domain.backup.RestoreRepo
-import com.shub39.grit.core.domain.backup.RestoreResult
-import com.shub39.grit.core.domain.backup.SchemaMismatchException
 import com.shub39.grit.core.habits.domain.HabitRepo
+import com.shub39.grit.core.settings.domain.backup.RestoreFailedException
+import com.shub39.grit.core.settings.domain.backup.RestoreRepo
+import com.shub39.grit.core.settings.domain.backup.RestoreResult
+import com.shub39.grit.core.settings.domain.backup.SchemaMismatchException
 import com.shub39.grit.core.tasks.domain.TaskRepo
 import com.shub39.grit.habits.data.database.HabitDatabase
 import com.shub39.grit.tasks.data.database.TaskDatabase
