+++ app/src/main/java/com/shub39/grit/viewmodels/SettingsViewModel.kt
@@ -24,14 +24,14 @@ import com.shub39.grit.core.data.BiometricUtilsImpl
 import com.shub39.grit.core.data.ChangelogManager
 import com.shub39.grit.core.domain.SettingsDatastore
 import com.shub39.grit.core.domain.ThemeDatastore
-import com.shub39.grit.core.domain.backup.ExportRepo
-import com.shub39.grit.core.domain.backup.ExportState
-import com.shub39.grit.core.domain.backup.RestoreRepo
-import com.shub39.grit.core.domain.backup.RestoreResult
-import com.shub39.grit.core.domain.backup.RestoreState
-import com.shub39.grit.core.presentation.settings.BackupState
-import com.shub39.grit.core.presentation.settings.SettingsAction
-import com.shub39.grit.core.presentation.settings.SettingsState
+import com.shub39.grit.core.settings.domain.backup.ExportRepo
+import com.shub39.grit.core.settings.domain.backup.ExportState
+import com.shub39.grit.core.settings.domain.backup.RestoreRepo
+import com.shub39.grit.core.settings.domain.backup.RestoreResult
+import com.shub39.grit.core.settings.domain.backup.RestoreState
+import com.shub39.grit.core.settings.presentation.BackupState
+import com.shub39.grit.core.settings.presentation.SettingsAction
+import com.shub39.grit.core.settings.presentation.SettingsState
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
