+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/settings/domain/backup/ExportRepo.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.domain.backup
+package com.shub39.grit.core.settings.domain.backup
 
 interface ExportRepo {
     suspend fun exportToJson()
