+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/settings/presentation/SettingsAction.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.presentation.settings
+package com.shub39.grit.core.settings.presentation
 
 import androidx.compose.ui.graphics.Color
 import com.shub39.grit.core.settings.domain.Sections
